% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbind.fill.R
\name{cbind.fill}
\alias{cbind.fill}
\title{cbind.fill}
\usage{
\method{cbind}{fill}(df_list)
}
\arguments{
\item{df_list}{a list of data.frames (required)}
}
\value{
a single data.frame where empty rows are filled with NAs
}
\description{
cbind a list of data.frames of same row names but unequal number.
}
