% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum.R
\name{dimsum}
\alias{dimsum}
\title{Run DiMSum pipeline}
\usage{
dimsum(
  runDemo = F,
  fastqFileDir = NULL,
  fastqFileExtension = ".fastq",
  gzipped = T,
  stranded = T,
  paired = T,
  barcodeDesignPath = NULL,
  barcodeErrorRate = 0.25,
  experimentDesignPath = NULL,
  experimentDesignPairDuplicates = F,
  barcodeIdentityPath = NULL,
  countPath = NULL,
  synonymSequencePath = NULL,
  cutadaptCut5First = NULL,
  cutadaptCut5Second = NULL,
  cutadaptCut3First = NULL,
  cutadaptCut3Second = NULL,
  cutadapt5First = NULL,
  cutadapt5Second = NULL,
  cutadapt3First = NULL,
  cutadapt3Second = NULL,
  cutadaptMinLength = 50,
  cutadaptErrorRate = 0.2,
  cutadaptOverlap = 3,
  vsearchMinQual = 30,
  vsearchMaxQual = 41,
  vsearchMaxee = 0.5,
  vsearchMinovlen = 10,
  outputPath = "./",
  projectName = "DiMSum_Project",
  wildtypeSequence = NULL,
  permittedSequences = NULL,
  reverseComplement = F,
  sequenceType = "auto",
  mutagenesisType = "random",
  transLibrary = F,
  transLibraryReverseComplement = F,
  bayesianDoubleFitness = F,
  bayesianDoubleFitnessLamD = 0.025,
  fitnessMinInputCountAll = "0",
  fitnessMinInputCountAny = "0",
  fitnessMinOutputCountAll = "0",
  fitnessMinOutputCountAny = "0",
  fitnessHighConfidenceCount = 10,
  fitnessDoubleHighConfidenceCount = 50,
  fitnessNormalise = T,
  fitnessErrorModel = T,
  fitnessDropoutPseudocount = 0,
  indels = "none",
  maxSubstitutions = 2,
  mixedSubstitutions = F,
  retainIntermediateFiles = F,
  splitChunkSize = 3758096384,
  retainedReplicates = "all",
  startStage = 0,
  stopStage = 5,
  numCores = 1
)
}
\arguments{
\item{runDemo}{Run the DiMSum demo (default:F)}

\item{fastqFileDir}{Path to directory containing input FASTQ files (required for WRAP)}

\item{fastqFileExtension}{FASTQ file extension (default:'.fastq')}

\item{gzipped}{Are FASTQ files are gzipped? (default:T)}

\item{stranded}{Is the library design stranded? (default:T)}

\item{paired}{Is the library design paired-end? (default:T)}

\item{barcodeDesignPath}{Path to barcode design file (tab-separated plain text file with barcode design)}

\item{barcodeErrorRate}{Maximum allowed error rate for barcode to be matched (default:0.25)}

\item{experimentDesignPath}{Path to Experimental Design File (required if '--runDemo'=F)}

\item{experimentDesignPairDuplicates}{Are multiple instances of FASTQ files in the Experimental Design File permitted? (default:F)}

\item{barcodeIdentityPath}{Path to Variant Identity File (tab-separated plain text file mapping barcodes to variants)}

\item{countPath}{Path to Variant Count File for analysis with STEAM only (tab-separated plain text file with sample counts for all variants)}

\item{synonymSequencePath}{Path to Synonym Sequences File with coding sequences for which synonymous variant fitness should be quantified (default: plain text file with one coding nucleotide sequence per line)}

\item{cutadaptCut5First}{Remove fixed number of bases from start (5') of first (or only) read before constant region trimming (optional)}

\item{cutadaptCut5Second}{Remove fixed number of bases from start (5') of second read in pair before constant region trimming (optional)}

\item{cutadaptCut3First}{Remove fixed number of bases from end (3') of first (or only) read before constant region trimming (optional)}

\item{cutadaptCut3Second}{Remove fixed number of bases from end (3') of second read in pair before constant region trimming (optional)}

\item{cutadapt5First}{Sequence of 5' constant region to be trimmed from first (or only) read (optional)}

\item{cutadapt5Second}{Sequence of 5' constant region to be trimmed from second read in pair (optional)}

\item{cutadapt3First}{Sequence of 3' constant region to be trimmed from first (or only) read (default: reverse complement of '--cutadapt5Second')}

\item{cutadapt3Second}{Sequence of 3' constant region to be trimmed from second read in pair (default: reverse complement of '--cutadapt5First')}

\item{cutadaptMinLength}{Discard reads shorter than LENGTH after trimming (default:50)}

\item{cutadaptErrorRate}{Maximum allowed error rate for trimming constant regions (default:0.2)}

\item{cutadaptOverlap}{Minimum overlap between read and constant region for trimming (default:3)}

\item{vsearchMinQual}{Minimum Phred base quality score required to retain read or read pair (default:30)}

\item{vsearchMaxQual}{Maximum Phred base quality score accepted when reading (and used when writing) FASTQ files; cannot be greater than 93 (default:41)}

\item{vsearchMaxee}{Maximum number of expected errors tolerated to retain read or read pair (default:0.5)}

\item{vsearchMinovlen}{Discard read pair if the alignment length is shorter than this (default:10)}

\item{outputPath}{Path to directory to use for output files (default:'./' i.e. current working directory)}

\item{projectName}{Project name and directory where results are to be saved (default:'DiMSum_Project')}

\item{wildtypeSequence}{Wild-type nucleotide sequence (A/C/G/T). Lower-case bases (a/c/g/t) indicate internal constant regions to be removed (required if '--runDemo'=F)}

\item{permittedSequences}{Nucleotide sequence of IUPAC ambiguity codes (A/C/G/T/R/Y/S/W/K/M/B/D/H/V/N) with length matching the number of mutated positions (i.e upper-case letters) in '--wildtypeSequence' (default:N i.e. any substitution mutation allowed)}

\item{reverseComplement}{Reverse complement sequence (default:F)}

\item{sequenceType}{Coding potential of sequence: either 'noncoding', 'coding' or 'auto'. If the specified wild-type nucleotide sequence ('--wildtypeSequence') has a valid translation without a premature STOP codon, it is assumed to be 'coding' (default:'auto')}

\item{mutagenesisType}{Whether mutagenesis was performed at the nucleotide or codon/amino acid level; either 'random' or 'codon' (default:'random')}

\item{transLibrary}{Paired-end reads correspond to distinct molecules? (default:F)}

\item{transLibraryReverseComplement}{Reverse complement second read in pair (default:F)}

\item{bayesianDoubleFitness}{In development: improve double mutant fitness estimates using Bayesian framework (DISABLED: still in development)}

\item{bayesianDoubleFitnessLamD}{In development: Poisson distribution for score likelihood (default:0.025)}

\item{fitnessMinInputCountAll}{Minimum input read count (in all replicates) to be retained during fitness calculations (default:0)}

\item{fitnessMinInputCountAny}{Minimum input read count (in any replicate) to be retained during fitness calculations (default:0)}

\item{fitnessMinOutputCountAll}{Minimum output read count (in all replicates) to be retained during fitness calculations (default:0)}

\item{fitnessMinOutputCountAny}{Minimum output read count (in any replicates) to be retained during fitness calculations (default:0)}

\item{fitnessHighConfidenceCount}{In development: minimum mean input read count for high confidence variants (default:10)}

\item{fitnessDoubleHighConfidenceCount}{In development: minimum input replicate read count for doubles used to derive prior for Bayesian doubles correction (default:50)}

\item{fitnessNormalise}{Normalise fitness values to minimise inter-replicate differences (default:T)}

\item{fitnessErrorModel}{Fit fitness error model (default:T)}

\item{fitnessDropoutPseudocount}{Pseudocount added to output replicates with dropout i.e. variants present in input but absent from output (default:0)}

\item{indels}{Indel variants to be retained: either 'all', 'none' or a comma-separated list of sequence lengths (default:'none')}

\item{maxSubstitutions}{Maximum number of nucleotide or amino acid substitutions for coding or non-coding sequences respectively (default:2)}

\item{mixedSubstitutions}{For coding sequences, are nonsynonymous variants with silent/synonymous substitutions in other codons allowed? (default:F)}

\item{retainIntermediateFiles}{Should intermediate files be retained? Intermediate files can be many gigabytes, but are required to rerun DiMSum starting at intermediate pipeline stages (default:F)}

\item{splitChunkSize}{Internal: FASTQ file split chunk size in bytes (default:3758096384)}

\item{retainedReplicates}{Comma-separated list of (integer) experiment replicates to retain or 'all' (default:'all')}

\item{startStage}{(Re-)Start DiMSum at a specific pipeline stage (default:0)}

\item{stopStage}{Stop DiMSum at a specific pipeline stage (default:5)}

\item{numCores}{Number of available CPU cores. All pipeline stages make use of parallel computing to decrease runtime if multiple cores are available (default:1)}
}
\value{
Nothing
}
\description{
This function runs the DiMSum pipeline.
}
