% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__add_dropout_pseudocount.R
\name{dimsum__add_dropout_pseudocount}
\alias{dimsum__add_dropout_pseudocount}
\title{dimsum__add_dropout_pseudocount}
\usage{
dimsum__add_dropout_pseudocount(dimsum_meta, input_dt, all_reps, verbose = T)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}

\item{input_dt}{input data.table (required)}

\item{all_reps}{list of replicates to retain (required)}

\item{verbose}{whether or not to print status messages (default: TRUE)}
}
\value{
data.table with pseudocounts added to output samples with dropout
}
\description{
Add pseudocount to output samples with dropout (zero read counts).
}
