% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__aggregate_AA_variants.R
\name{dimsum__aggregate_AA_variants}
\alias{dimsum__aggregate_AA_variants}
\title{dimsum__aggregate_AA_variants}
\usage{
dimsum__aggregate_AA_variants(input_dt)
}
\arguments{
\item{input_dt}{output path for plots and saved objects (required)}
}
\value{
Nothing
}
\description{
Aggregate counts from variants that are identical at the AA level and without synonymous mutations.
}
