% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__aggregate_AA_variants_fitness.R
\name{dimsum__aggregate_AA_variants_fitness}
\alias{dimsum__aggregate_AA_variants_fitness}
\title{dimsum__aggregate_AA_variants_fitness}
\usage{
dimsum__aggregate_AA_variants_fitness(dimsum_meta, input_dt, all_reps)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}

\item{input_dt}{input data.table (required)}

\item{all_reps}{list of replicates to retain (required)}
}
\value{
data.table with aggregated counts
}
\description{
Aggregate fitness and error from variants that are identical at the AA level and without synonymous mutations.
}
