% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__calculate_fitness.R
\name{dimsum__calculate_fitness}
\alias{dimsum__calculate_fitness}
\title{dimsum__calculate_fitness}
\usage{
dimsum__calculate_fitness(
  dimsum_meta,
  input_dt,
  all_reps,
  error_model_dt,
  norm_model_dt,
  verbose = T
)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}

\item{input_dt}{input data.table (required)}

\item{all_reps}{list of replicates to retain (required)}

\item{error_model_dt}{error model data.table (required)}

\item{norm_model_dt}{normalisation model data.table (required)}

\item{verbose}{whether or not to print status messages (default: TRUE)}
}
\value{
data.table with fitness and error
}
\description{
Calculate fitness and error.
}
