% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__cbind_fill.R
\name{dimsum__cbind_fill}
\alias{dimsum__cbind_fill}
\title{dimsum__cbind_fill}
\usage{
dimsum__cbind_fill(df_list)
}
\arguments{
\item{df_list}{a list of data.frames (required)}
}
\value{
a single data.frame where empty rows are filled with NAs
}
\description{
cbind a list of data.frames of same row names but unequal number.
}
