% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__check_barcode_design.R
\name{dimsum__check_barcode_design}
\alias{dimsum__check_barcode_design}
\title{dimsum__check_barcode_design}
\usage{
dimsum__check_barcode_design(barcode_design)
}
\arguments{
\item{barcode_design}{barcode design data.frame (required)}
}
\value{
Nothing
}
\description{
Validate metadata from barcode design file.
}
