% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__check_barcodeidentityfile.R
\name{dimsum__check_barcodeidentityfile}
\alias{dimsum__check_barcodeidentityfile}
\title{dimsum__check_barcodeidentityfile}
\usage{
dimsum__check_barcodeidentityfile(dimsum_meta, return_data = FALSE)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}

\item{return_data}{whether or not to return data (default: FALSE)}
}
\value{
Reformatted data.table
}
\description{
Check whether user-specified barcode identity file correctly formatted.
}
