% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__check_experiment_design_countfile.R
\name{dimsum__check_experiment_design_countfile}
\alias{dimsum__check_experiment_design_countfile}
\title{dimsum__check_experiment_design_countfile}
\usage{
dimsum__check_experiment_design_countfile(dimsum_meta, exp_design)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}

\item{exp_design}{experiment design data.frame (required)}
}
\value{
Nothing
}
\description{
Validate metadata from experiment design file when countPath specified.
}
