% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__check_files_exist.R
\name{dimsum__check_files_exist}
\alias{dimsum__check_files_exist}
\title{dimsum__check_files_exist}
\usage{
dimsum__check_files_exist(
  required_files,
  stage_number = NULL,
  execute = TRUE,
  exit = TRUE
)
}
\arguments{
\item{required_files}{character vector of file paths (required)}

\item{stage_number}{integer stage number (required)}

\item{execute}{whether this stage is to be executed (default:TRUE)}

\item{exit}{whether to stop execution (default:TRUE)}
}
\value{
Nothing
}
\description{
Check files exist and if not produce error message and stop.
}
