% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__check_nt_fitness.R
\name{dimsum__check_nt_fitness}
\alias{dimsum__check_nt_fitness}
\title{dimsum__check_nt_fitness}
\usage{
dimsum__check_nt_fitness(input_dt, all_reps, report = TRUE,
  report_outpath = NULL)
}
\arguments{
\item{input_dt}{input data.table (required)}

\item{all_reps}{list of replicates to retain (required)}

\item{report}{whether or not to generate fitness summary plots (default: TRUE)}

\item{report_outpath}{fitness report output path}
}
\value{
Nothing
}
\description{
Check fitness and count-based error of single and double nucleotide substitution variants.
}
