% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__check_variants.R
\name{dimsum__check_variants}
\alias{dimsum__check_variants}
\title{dimsum__check_variants}
\usage{
dimsum__check_variants(dimsum_meta, input_dt)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}

\item{input_dt}{input data.table (required)}
}
\value{
Nothing
}
\description{
Check whether minimum variant requirements satisfied.
}
