% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__convert_linked_adapters.R
\name{dimsum__convert_linked_adapters}
\alias{dimsum__convert_linked_adapters}
\title{dimsum__convert_linked_adapters}
\usage{
dimsum__convert_linked_adapters(dimsum_meta)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}
}
\value{
An updated dimsum_meta object
}
\description{
Convert to linked adapters if 3' constant region expected to be sequenced.
}
