% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__create_dir.R
\name{dimsum__create_dir}
\alias{dimsum__create_dir}
\title{dimsum__create_dir}
\usage{
dimsum__create_dir(dimsum_dir, execute, message = NULL,
  overwrite_dir = TRUE)
}
\arguments{
\item{dimsum_dir}{directory path string (required)}

\item{execute}{whether or not the system command will be executed (required)}

\item{message}{message string (optional, default: NULL i.e. no message displayed)}

\item{overwrite_dir}{delete directory if already exists (optional, default: TRUE)}
}
\value{
Nothing
}
\description{
Create results folder for dimsum pipeline stage.
}
