% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__cutadapt_report.R
\name{dimsum__cutadapt_report}
\alias{dimsum__cutadapt_report}
\title{dimsum__cutadapt_report}
\usage{
dimsum__cutadapt_report(dimsum_meta, report_outpath)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}

\item{report_outpath}{cutadapt report output path (required)}
}
\value{
an updated experiment metadata object
}
\description{
Generate cutadapt summary plots for all fastq files.
}
