% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__debarcode_variants.R
\name{dimsum__debarcode_variants}
\alias{dimsum__debarcode_variants}
\title{dimsum__debarcode_variants}
\usage{
dimsum__debarcode_variants(input_dt, barcode_path)
}
\arguments{
\item{input_dt}{input data.table (required)}

\item{barcode_path}{path to barcode identity file (required)}
}
\value{
A data.table with variants debarcoded
}
\description{
Debarcode variants.
}
