% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__demultiplex_cp_helper.R
\name{dimsum__demultiplex_cp_helper}
\alias{dimsum__demultiplex_cp_helper}
\title{dimsum__demultiplex_cp_helper}
\usage{
dimsum__demultiplex_cp_helper(i)
}
\arguments{
\item{i}{"fastq_pair_list" row index (required)}
}
\value{
Nothing
}
\description{
Helper function to copy and rename fastq files before demultiplexing. "dimsum_meta","fastq_pair_list" and "demultiplex_outpath" objects need to be available globally.
}
