% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__demultiplex_helper.R
\name{dimsum__demultiplex_helper}
\alias{dimsum__demultiplex_helper}
\title{dimsum__demultiplex_helper}
\usage{
dimsum__demultiplex_helper(i)
}
\arguments{
\item{i}{"fastq_pair_list" row index (required)}
}
\value{
Nothing
}
\description{
Helper function to demultiplex fastq files. "dimsum_meta", "demultiplex_outpath" and "fastq_pair_list" objects need to be available globally.
}
