% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__derange_df.R
\name{dimsum__derange_df}
\alias{dimsum__derange_df}
\title{dimsum__derange_df}
\usage{
dimsum__derange_df(input_df)
}
\arguments{
\item{input_df}{input data.frame (required)}
}
\value{
a single deranged data.frame
}
\description{
Expand a data.frame with ranged row.names.
}
