% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__detect_sequence_type.R
\name{dimsum__detect_sequence_type}
\alias{dimsum__detect_sequence_type}
\title{dimsum__detect_sequence_type}
\usage{
dimsum__detect_sequence_type(input_sequence)
}
\arguments{
\item{input_sequence}{input nucleotide sequence (required)}
}
\value{
Nothing
}
\description{
Identify and annotate double AA substitutions.
}
