% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__error_model.R
\name{dimsum__error_model}
\alias{dimsum__error_model}
\title{dimsum__error_model}
\usage{
dimsum__error_model(
  dimsum_meta,
  input_dt,
  all_reps,
  report = TRUE,
  report_outpath = NULL
)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}

\item{input_dt}{input data.table (required)}

\item{all_reps}{list of replicates to retain (required)}

\item{report}{whether or not to generate fitness summary plots (default: TRUE)}

\item{report_outpath}{fitness report output path}
}
\value{
error model data.table
}
\description{
Error model analysis including determination of count-based, replicate and over-sequencing error terms.
}
