% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__error_model_qqplot.R
\name{dimsum__error_model_qqplot}
\alias{dimsum__error_model_qqplot}
\title{dimsum__error_model_qqplot}
\usage{
dimsum__error_model_qqplot(
  dimsum_meta,
  input_dt,
  all_reps,
  norm_dt,
  error_dt,
  report_outpath = NULL
)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}

\item{input_dt}{input data.table (required)}

\item{all_reps}{list of replicates to retain (required)}

\item{norm_dt}{data.table of normalisation parameters (required)}

\item{error_dt}{data.table of error model parameters (required)}

\item{report_outpath}{fitness report output path (default:NULL)}
}
\value{
Nothing
}
\description{
Perform leave one out cross validation on replicates and generate QQ plot
}
