% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__filter_nuc_variants_noncoding.R
\name{dimsum__filter_nuc_variants_noncoding}
\alias{dimsum__filter_nuc_variants_noncoding}
\title{dimsum__filter_nuc_variants_noncoding}
\usage{
dimsum__filter_nuc_variants_noncoding(dimsum_meta, input_dt, wt_ntseq,
  all_reps, report = TRUE, report_outpath = NULL)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}

\item{input_dt}{output path for plots and saved objects (required)}

\item{wt_ntseq}{WT nucleotide sequence (required)}

\item{all_reps}{list of replicates to retain (required)}

\item{report}{whether or not to generate fitness summary plots (default: TRUE)}

\item{report_outpath}{fitness report output path}
}
\value{
Nothing
}
\description{
Filter nucleotide variants from noncoding sequences.
}
