% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__filter_reads.R
\name{dimsum__filter_reads}
\alias{dimsum__filter_reads}
\title{dimsum__filter_reads}
\usage{
dimsum__filter_reads(
  input_FASTQ,
  input_REPORT,
  output_FASTQ,
  output_REPORT,
  min_qual
)
}
\arguments{
\item{input_FASTQ}{Path to input FASTQ file (required)}

\item{input_REPORT}{Path to input report file (required)}

\item{output_FASTQ}{Path to output FASTQ file (required)}

\item{output_REPORT}{Path to output report file (required)}

\item{min_qual}{Minimum observed base quality to retain read pair (required)}
}
\value{
Nothing
}
\description{
Concatenate reads (with or without reverse complementing second read in pair).
}
