% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__filter_reads_helper.R
\name{dimsum__filter_reads_helper}
\alias{dimsum__filter_reads_helper}
\title{dimsum__filter_reads_helper}
\usage{
dimsum__filter_reads_helper(i)
}
\arguments{
\item{i}{"sample_names" index (required)}
}
\value{
Nothing
}
\description{
Helper function to filter reads. "dimsum_meta", "vsearch_outpath", "sample_names" and "dimsum__filter_reads" objects need to be available globally.
}
