% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__filter_single_end_reads.R
\name{dimsum__filter_single_end_reads}
\alias{dimsum__filter_single_end_reads}
\title{dimsum__filter_single_end_reads}
\usage{
dimsum__filter_single_end_reads(input_FASTQ, output_FASTQ, output_REPORT,
  min_qual, max_ee, min_len)
}
\arguments{
\item{input_FASTQ}{Path to first read FASTQ file (required)}

\item{output_FASTQ}{Path to output FASTQ file (required)}

\item{output_REPORT}{Path to report file (required)}

\item{min_qual}{Minimum observed base quality to retain read pair (required)}

\item{max_ee}{Maximum number of expected errors to retain read pair (required)}

\item{min_len}{Discard pair if either read is shorter than this (required)}
}
\value{
Nothing
}
\description{
Concatentate reads (without reverse complementing second read in pair).
}
