% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__filter_variants_noncoding.R
\name{dimsum__filter_variants_noncoding}
\alias{dimsum__filter_variants_noncoding}
\title{dimsum__filter_variants_noncoding}
\usage{
dimsum__filter_variants_noncoding(dimsum_meta, input_dt, wt_ntseq, all_reps)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}

\item{input_dt}{input data.table (required)}

\item{wt_ntseq}{WT nucleotide sequence (required)}

\item{all_reps}{list of replicates to retain (required)}
}
\value{
Nothing
}
\description{
Filter for desired non-coding variants.
}
