% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__fit_error_model.R
\name{dimsum__fit_error_model}
\alias{dimsum__fit_error_model}
\title{dimsum__fit_error_model}
\usage{
dimsum__fit_error_model(dimsum_meta, input_dt, all_reps, norm_dt,
  Nbootstraps = 100, maxN = 10000, max_tries_per_fit = 20,
  lower_rep = 1e-04)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}

\item{input_dt}{input data.table (required)}

\item{all_reps}{list of replicates to retain (required)}

\item{norm_dt}{data.table of normalisation parameters (default:NULL)}

\item{Nbootstraps}{number of bootstraps (default:100)}

\item{maxN}{maximum sample size for each bootstrap (default:5000)}

\item{max_tries_per_fit}{maximum number of fitting attempts (default:20)}

\item{lower_rep}{replicate error lower bound (default:1e-4)}
}
\value{
Nothing
}
\description{
Fit error model.
}
