% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__fit_error_model_bootstrap.R
\name{dimsum__fit_error_model_bootstrap}
\alias{dimsum__fit_error_model_bootstrap}
\title{dimsum__fit_error_model_bootstrap}
\usage{
dimsum__fit_error_model_bootstrap(
  m,
  input_dt,
  all_reps,
  idx,
  maxN,
  max_tries_per_fit,
  lower_rep,
  Fcorr = NULL
)
}
\arguments{
\item{m}{bootstrap number; dummy variable (required)}

\item{input_dt}{input data.table (required)}

\item{all_reps}{list of replicates to retain (required)}

\item{idx}{vector of replicate combinations (required)}

\item{maxN}{maximum size of subsample (required)}

\item{max_tries_per_fit}{maximum number of fitting attempts (required)}

\item{lower_rep}{replicate error lower bound (required)}

\item{Fcorr}{vector of fitness normalization factors (default:NULL)}
}
\value{
Nothing
}
\description{
Fit error model on random subsample of the input data.
}
