% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__get_barcode_design.R
\name{dimsum__get_barcode_design}
\alias{dimsum__get_barcode_design}
\title{dimsum__get_barcode_design}
\usage{
dimsum__get_barcode_design(dimsum_meta)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}
}
\value{
a data.frame with the validated barcode design
}
\description{
Get, format and validate metadata from barcode design file.
}
