% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__get_cutadapt_options.R
\name{dimsum__get_cutadapt_options}
\alias{dimsum__get_cutadapt_options}
\title{dimsum__get_cutadapt_options}
\usage{
dimsum__get_cutadapt_options(dimsum_meta, exp_design_row,
  option_type = "default")
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}

\item{exp_design_row}{row number of exp_design table (required)}

\item{option_type}{one of default/cut/swap (default:"default")}
}
\value{
A string of command-line options for cutadapt
}
\description{
Get command-line options for cutadapt.
}
