% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__get_synonymsequences.R
\name{dimsum__get_synonymsequences}
\alias{dimsum__get_synonymsequences}
\title{dimsum__get_synonymsequences}
\usage{
dimsum__get_synonymsequences(dimsum_meta)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}
}
\value{
character vector of sequences
}
\description{
Get and check whether user-specified synonym sequences correctly formatted.
}
