% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__gg_color_hue.R
\name{dimsum__gg_color_hue}
\alias{dimsum__gg_color_hue}
\title{dimsum__gg_color_hue}
\usage{
dimsum__gg_color_hue(n)
}
\arguments{
\item{n}{an integer number of colours required (required)}
}
\value{
a vector of colours
}
\description{
ggplot2-like colour scale in HCL space.
}
