% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__ggpairs_binhex.R
\name{dimsum__ggpairs_binhex}
\alias{dimsum__ggpairs_binhex}
\title{dimsum__ggpairs_binhex}
\usage{
dimsum__ggpairs_binhex(input_dt, output_file, width = 12, height = 12,
  bins = 50, xlab = "x", ylab = "y", title = "")
}
\arguments{
\item{input_dt}{input data table (required)}

\item{output_file}{plot output path (required)}

\item{width}{plot width (default: 12)}

\item{height}{plot height (default: 12)}

\item{bins}{number of hexagon bins (default: 50)}

\item{xlab}{the title of the x-axis}

\item{ylab}{the title of the y-axis}

\item{title}{the plot title}
}
\value{
Nothing
}
\description{
GGpairs plot for all variables in input data table with correlation in upper triangle and 2d binned hexagons in lower triangle.
}
