% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__hamming_distance.R
\name{dimsum__hamming_distance}
\alias{dimsum__hamming_distance}
\title{dimsum__hamming_distance}
\usage{
dimsum__hamming_distance(x, y)
}
\arguments{
\item{x}{input string 1 (required)}

\item{y}{input string 2 (required)}
}
\value{
hamming distance
}
\description{
Calculate hamming distance of two strings of the same length.
}
