% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__identify_double_aa_mutations.R
\name{dimsum__identify_double_aa_mutations}
\alias{dimsum__identify_double_aa_mutations}
\title{dimsum__identify_double_aa_mutations}
\usage{
dimsum__identify_double_aa_mutations(input_dt, singles_dt, wt_AAseq,
  report = TRUE, report_outpath = NULL)
}
\arguments{
\item{input_dt}{input data.table (required)}

\item{singles_dt}{singles data.table (required)}

\item{wt_AAseq}{WT amino acid sequence (required)}

\item{report}{whether or not to generate fitness summary plots (default: TRUE)}

\item{report_outpath}{fitness report output path}
}
\value{
Nothing
}
\description{
Identify and annotate double AA substitutions.
}
