% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__identify_internal_constant_region.R
\name{dimsum__identify_internal_constant_region}
\alias{dimsum__identify_internal_constant_region}
\title{dimsum__identify_internal_constant_region}
\usage{
dimsum__identify_internal_constant_region(input_dt, wt_ccntseq)
}
\arguments{
\item{input_dt}{input data.table (required)}

\item{wt_ccntseq}{case-coded WT nucleotide sequence (required)}
}
\value{
A data.table with internal constant regions identified from nucleotide sequences
}
\description{
Identify internal constant region from nucleotide sequences.
}
