% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__identify_permitted_mutations.R
\name{dimsum__identify_permitted_mutations}
\alias{dimsum__identify_permitted_mutations}
\title{dimsum__identify_permitted_mutations}
\usage{
dimsum__identify_permitted_mutations(dimsum_meta, input_dt)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}

\item{input_dt}{input data.table (required)}
}
\value{
A data.table with permitted variants identified
}
\description{
Identify variants with permitted mutations.
}
