% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__identify_single_aa_mutations.R
\name{dimsum__identify_single_aa_mutations}
\alias{dimsum__identify_single_aa_mutations}
\title{dimsum__identify_single_aa_mutations}
\usage{
dimsum__identify_single_aa_mutations(input_dt, wt_AAseq)
}
\arguments{
\item{input_dt}{input data.table (required)}

\item{wt_AAseq}{WT amino acid sequence (required)}
}
\value{
data.table with single amino acid variants
}
\description{
Identify and annotate single AA substitutions.
}
