% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__identify_single_nt_mutations.R
\name{dimsum__identify_single_nt_mutations}
\alias{dimsum__identify_single_nt_mutations}
\title{dimsum__identify_single_nt_mutations}
\usage{
dimsum__identify_single_nt_mutations(input_dt, wt_ntseq)
}
\arguments{
\item{input_dt}{input data.table (required)}

\item{wt_ntseq}{WT nucleotide sequence (required)}
}
\value{
data.table with single nucleotide variants
}
\description{
Identify and annotate single nucleotide substitutions.
}
