% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__identify_STOP_mutations.R
\name{dimsum__identify_STOP_mutations}
\alias{dimsum__identify_STOP_mutations}
\title{dimsum__identify_STOP_mutations}
\usage{
dimsum__identify_STOP_mutations(input_dt)
}
\arguments{
\item{input_dt}{input data.table (required)}
}
\value{
data.table with single nucleotide variants
}
\description{
Identify and annotate single nucleotide substitutions.
}
