% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__infer_growth_rates.R
\name{dimsum__infer_growth_rates}
\alias{dimsum__infer_growth_rates}
\title{dimsum__infer_growth_rates}
\usage{
dimsum__infer_growth_rates(dimsum_meta, input_dt, all_reps)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}

\item{input_dt}{input data.table (required)}

\item{all_reps}{list of replicates to retain (required)}
}
\value{
data.table with growth rate and growth rate error
}
\description{
Infer growth rates from fitness.
}
