% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__merge_fitness.R
\name{dimsum__merge_fitness}
\alias{dimsum__merge_fitness}
\title{dimsum__merge_fitness}
\usage{
dimsum__merge_fitness(dimsum_meta, input_dt, doubles_dt, singles_dt,
  all_reps, fitness_outpath, report = TRUE, report_outpath = NULL)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}

\item{input_dt}{input data.table (required)}

\item{doubles_dt}{doubles data.table (required)}

\item{singles_dt}{singles data.table (required)}

\item{all_reps}{list of replicates to retain (required)}

\item{fitness_outpath}{output path for saved objects (required)}

\item{report}{whether or not to generate fitness summary plots (default: TRUE)}

\item{report_outpath}{fitness report output path}
}
\value{
Nothing
}
\description{
Calculate fitness and count-based error.
}
