% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__merge_report.R
\name{dimsum__merge_report}
\alias{dimsum__merge_report}
\title{dimsum__merge_report}
\usage{
dimsum__merge_report(dimsum_meta, report_outpath)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}

\item{report_outpath}{final report output path (required)}
}
\value{
a single data.frame where empty rows are filled with NAs
}
\description{
Generate final summary plots for all samples.
}
