% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__normalise_fitness.R
\name{dimsum__normalise_fitness}
\alias{dimsum__normalise_fitness}
\title{dimsum__normalise_fitness}
\usage{
dimsum__normalise_fitness(dimsum_meta, input_dt, all_reps,
  fitness_suffix = "")
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}

\item{input_dt}{input data.table (required)}

\item{all_reps}{list of replicates to retain (required)}

\item{fitness_suffix}{fitness and sigma suffix (default:"")}
}
\value{
Nothing
}
\description{
Normalise fitness and error for differences in number of generations.
}
