% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__parse_minimum_read_count_arguments.R
\name{dimsum__parse_minimum_read_count_arguments}
\alias{dimsum__parse_minimum_read_count_arguments}
\title{dimsum__parse_minimum_read_count_arguments}
\usage{
dimsum__parse_minimum_read_count_arguments(input_arg)
}
\arguments{
\item{input_arg}{a string (required)}
}
\value{
a single integer or list of named integers
}
\description{
Parse minimum input count arguments.
}
