% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__plot_samplename.R
\name{dimsum__plot_samplename}
\alias{dimsum__plot_samplename}
\title{dimsum__plot_samplename}
\usage{
dimsum__plot_samplename(input_names)
}
\arguments{
\item{input_names}{a list of sample names (required)}
}
\value{
reformatted sample names
}
\description{
Format sample names for plotting purposes.
}
