% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__process_merged_variants.R
\name{dimsum__process_merged_variants}
\alias{dimsum__process_merged_variants}
\title{dimsum__process_merged_variants}
\usage{
dimsum__process_merged_variants(dimsum_meta, input_dt)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}

\item{input_dt}{output path for plots and saved objects (required)}
}
\value{
A list of data.tables containing indel variants, rejected variants (internal constant region variants, not permitted sequences and variants with too many substititions) and retained variants
}
\description{
Process nucleotide variants to remove indels, internal constant region variants, not permitted sequences and variants with too many substitutions.
}
