% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__range_df.R
\name{dimsum__range_df}
\alias{dimsum__range_df}
\title{dimsum__range_df}
\usage{
dimsum__range_df(input_df)
}
\arguments{
\item{input_df}{input data.frame (required)}
}
\value{
a single ranged data.frame
}
\description{
Collapse a data.frame with integer rownames to ranged row.names.
}
