% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__remove_forbidden_mutations.R
\name{dimsum__remove_forbidden_mutations}
\alias{dimsum__remove_forbidden_mutations}
\title{dimsum__remove_forbidden_mutations}
\usage{
dimsum__remove_forbidden_mutations(dimsum_meta, input_dt)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}

\item{input_dt}{input data.table (required)}
}
\value{
A data.table with variants with permitted mutations only
}
\description{
Subset variants to those with permitted mutations.
}
