% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__render_report.R
\name{dimsum__render_report}
\alias{dimsum__render_report}
\title{dimsum__render_report}
\usage{
dimsum__render_report(dimsum_meta, initialise = FALSE, finalise = FALSE)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}

\item{initialise}{whether to initialise plots with touch (default:F)}

\item{finalise}{whether to finalse the report with end time (default:F)}
}
\value{
Nothing
}
\description{
Generate a summary report in html format.
}
