% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__replicate_fitness_deviation.R
\name{dimsum__replicate_fitness_deviation}
\alias{dimsum__replicate_fitness_deviation}
\title{dimsum__replicate_fitness_deviation}
\usage{
dimsum__replicate_fitness_deviation(p, fitness_mat, all_reps)
}
\arguments{
\item{p}{vector of normalisation parameters (required)}

\item{fitness_mat}{matrix of fitness values (required)}

\item{all_reps}{list of replicates to retain (required)}
}
\value{
the square-root of sum of squared differences to mean over replicates
}
\description{
Calculate (square-root of sum of squared) differences between replicate fitness values (for all variants) given normalisation parameters
}
