% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__reports_summary.R
\name{dimsum__reports_summary}
\alias{dimsum__reports_summary}
\title{dimsum__reports_summary}
\usage{
dimsum__reports_summary(dimsum_meta)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}
}
\value{
a summary report character string (html format)
}
\description{
Generate a summary report in html format.
}
