% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__sample_count_distributions.R
\name{dimsum__sample_count_distributions}
\alias{dimsum__sample_count_distributions}
\title{dimsum__sample_count_distributions}
\usage{
dimsum__sample_count_distributions(input_dt, output_file, width = 12,
  height = 8, title = "", error_rate = 0.001, seq_length)
}
\arguments{
\item{input_dt}{input data table with count, WT and remaning column used for facetting (required)}

\item{output_file}{plot output path (required)}

\item{width}{plot width (default: 12)}

\item{height}{plot height (default: 8)}

\item{title}{plot title (default: "")}

\item{error_rate}{sequencing error rate (default: 1e-3)}

\item{seq_length}{sequence length}
}
\value{
Nothing
}
\description{
Plot sample nucleotide count distributions split by hamming distance (Nham_nt)
}
