% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__sample_count_distributions_by_ntham.R
\name{dimsum__sample_count_distributions_by_ntham}
\alias{dimsum__sample_count_distributions_by_ntham}
\title{dimsum__sample_count_distributions_by_ntham}
\usage{
dimsum__sample_count_distributions_by_ntham(input_dt, sequence_type,
  output_file, width = 12, height = 8, title = "")
}
\arguments{
\item{input_dt}{input data table (required)}

\item{sequence_type}{coding potential of sequence: noncoding/coding (required)}

\item{output_file}{plot output path (required)}

\item{width}{plot width (default: 12)}

\item{height}{plot height (default: 8)}

\item{title}{plot title (default: "")}
}
\value{
Nothing
}
\description{
Plot sample sample count distributions split by nucleotide hamming distance (Nham_nt) and synonymous/nonsynonymous mutations (Nham_aa).
}
