% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__save_png.R
\name{dimsum__save_png}
\alias{dimsum__save_png}
\title{dimsum__save_png}
\usage{
dimsum__save_png(
  output_file,
  plot_object,
  width = 7,
  height = 7,
  plot_dpi = 300
)
}
\arguments{
\item{output_file}{output file path (required)}

\item{plot_object}{plot object (required)}

\item{width}{plot width in inches (default:7)}

\item{height}{plot height in inches (default:7)}

\item{plot_dpi}{dots per inch (default:300)}
}
\value{
Nothing
}
\description{
Save plot object as PNG.
}
