% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__split_helper.R
\name{dimsum__split_helper}
\alias{dimsum__split_helper}
\title{dimsum__split_helper}
\usage{
dimsum__split_helper(i)
}
\arguments{
\item{i}{"fastq_pair_list" row index (required)}
}
\value{
Nothing
}
\description{
Helper function to split fastq files. "dimsum_meta", "fastq_pair_list", "split_outpath", "dimsum__fastq_splitter" and "dimsum__writeFastq" objects need to be available globally.
}
