% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__status_message.R
\name{dimsum__status_message}
\alias{dimsum__status_message}
\title{dimsum__status_message}
\usage{
dimsum__status_message(...)
}
\arguments{
\item{...}{objects to be output as character vector (required)}
}
\value{
Nothing
}
\description{
Write status message to standard output.
}
