% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__sum_datatable_columns.R
\name{dimsum__sum_datatable_columns}
\alias{dimsum__sum_datatable_columns}
\title{dimsum__sum_datatable_columns}
\usage{
dimsum__sum_datatable_columns(dt, column_patterns, suffix = "")
}
\arguments{
\item{dt}{Input data.table (required)}

\item{column_patterns}{character vector of column patterns to match (required)}

\item{suffix}{a character suffix for the appended column (default: "")}
}
\value{
a data.table where a subset of columns is replaced with a single column containing the row sums
}
\description{
Replace a subset of columns with a single column containing the row sums.
}
