% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__unique_helper.R
\name{dimsum__unique_helper}
\alias{dimsum__unique_helper}
\title{dimsum__unique_helper}
\usage{
dimsum__unique_helper(i)
}
\arguments{
\item{i}{"sample_names" index (required)}
}
\value{
Nothing
}
\description{
Helper function to concatenate split files. "dimsum_meta" and "unique_outpath" objects need to be available globally.
}
