% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__unzip_helper.R
\name{dimsum__unzip_helper}
\alias{dimsum__unzip_helper}
\title{dimsum__unzip_helper}
\usage{
dimsum__unzip_helper(i)
}
\arguments{
\item{i}{"all_fastq" index (required)}
}
\value{
Nothing
}
\description{
Helper function to unzip fastq files. "all_fastq" and "fastq_outpath" objects need to be available globally.
}
