% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__validate_input.R
\name{dimsum__validate_input}
\alias{dimsum__validate_input}
\title{dimsum__validate_input}
\usage{
dimsum__validate_input(input_list)
}
\arguments{
\item{input_list}{a list of input arguments (required)}
}
\value{
an experiment metadata object
}
\description{
Check validity and reformat input parameters.
}
