% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__vsearch_report.R
\name{dimsum__vsearch_report}
\alias{dimsum__vsearch_report}
\title{dimsum__vsearch_report}
\usage{
dimsum__vsearch_report(dimsum_meta, report_outpath)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}

\item{report_outpath}{VSEARCH report output path (required)}
}
\value{
an updated experiment metadata object
}
\description{
Generate VSEARCH summary plots for all samples.
}
