% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__vsearch_single_end_library_helper.R
\name{dimsum__vsearch_single_end_library_helper}
\alias{dimsum__vsearch_single_end_library_helper}
\title{dimsum__vsearch_single_end_library_helper}
\usage{
dimsum__vsearch_single_end_library_helper(i)
}
\arguments{
\item{i}{"sample_names" index (required)}
}
\value{
Nothing
}
\description{
Helper function to filter single end reads. "dimsum_meta", "vsearch_outpath", "sample_names" and "dimsum__filter_single_end_reads" objects need to be available globally.
}
