% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__vsearch_trans_library_helper.R
\name{dimsum__vsearch_trans_library_helper}
\alias{dimsum__vsearch_trans_library_helper}
\title{dimsum__vsearch_trans_library_helper}
\usage{
dimsum__vsearch_trans_library_helper(i)
}
\arguments{
\item{i}{"sample_names" index (required)}
}
\value{
Nothing
}
\description{
Helper function to merge trans library reads. "dimsum_meta", "vsearch_outpath", "sample_names" and "dimsum__concatenate_reads" objects need to be available globally.
}
