% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__writeFastq.R
\name{dimsum__writeFastq}
\alias{dimsum__writeFastq}
\title{dimsum__writeFastq}
\usage{
dimsum__writeFastq(shortreads, outputFile, initial_write)
}
\arguments{
\item{shortreads}{ShortReadQ object (required)}

\item{outputFile}{Path to output FASTQ file (required)}

\item{initial_write}{Is this the first write? (required)}
}
\value{
Nothing
}
\description{
writeFastq wrapper.
}
