% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum_stage_counts_to_fitness.R
\name{dimsum_stage_counts_to_fitness}
\alias{dimsum_stage_counts_to_fitness}
\title{dimsum_stage_counts_to_fitness}
\usage{
dimsum_stage_counts_to_fitness(dimsum_meta, fitness_outpath,
  all_reps = "all", min_mean_input_read_count = 10,
  min_input_read_count_doubles = 50, lam_d = 0.025, execute = TRUE,
  report = TRUE, report_outpath = NULL, save_workspace = TRUE)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}

\item{fitness_outpath}{output path for saved objects (required)}

\item{all_reps}{list of replicates to retain or "all" (default:"all")}

\item{min_mean_input_read_count}{minimum mean input read count for high confidence variants (default:10)}

\item{min_input_read_count_doubles}{minimum input read count for doubles used to derive prior for Bayesian doubles correction (default:50)}

\item{lam_d}{Poisson distribution for score likelihood (default:0.025)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}

\item{report}{whether or not to generate fitness summary plots (default: TRUE)}

\item{report_outpath}{fitness report output path}

\item{save_workspace}{whether or not to save the current workspace (default: TRUE)}
}
\value{
Nothing
}
\description{
Estimate fitness of singles, doubles and silent AA substitution variants from DiMSum formatted output counts.
}
