% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum_stage_cutadapt.R
\name{dimsum_stage_cutadapt}
\alias{dimsum_stage_cutadapt}
\title{dimsum_stage_cutadapt}
\usage{
dimsum_stage_cutadapt(
  dimsum_meta,
  cutadapt_outpath,
  report = TRUE,
  report_outpath = NULL,
  save_workspace = TRUE
)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}

\item{cutadapt_outpath}{cutadapt output path (required)}

\item{report}{whether or not to generate cutadapt summary plots (default: TRUE)}

\item{report_outpath}{cutadapt report output path}

\item{save_workspace}{whether or not to save the current workspace (default: TRUE)}
}
\value{
an updated experiment metadata object
}
\description{
Run cutadapt on all fastq files.
}
