% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum_stage_demultiplex.R
\name{dimsum_stage_demultiplex}
\alias{dimsum_stage_demultiplex}
\title{dimsum_stage_demultiplex}
\usage{
dimsum_stage_demultiplex(dimsum_meta, demultiplex_outpath,
  save_workspace = TRUE)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}

\item{demultiplex_outpath}{demultiplex output path (required)}

\item{save_workspace}{whether or not to save the current workspace (default: TRUE)}
}
\value{
an updated experiment metadata object
}
\description{
Run demultiplexing on all fastq files using cutadapt.
}
