% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum_stage_diagnostics_report.R
\name{dimsum_stage_diagnostics_report}
\alias{dimsum_stage_diagnostics_report}
\title{dimsum_stage_diagnostics_report}
\usage{
dimsum_stage_diagnostics_report(variant_data, report_outpath,
  input_samples_pattern = "^input", output_samples_pattern = "^output")
}
\arguments{
\item{variant_data}{path to variant data R Data file (required)}

\item{report_outpath}{diagnostics report output path (required)}

\item{input_samples_pattern}{input samples string pattern (default: "^input")}

\item{output_samples_pattern}{diagnostics report output path (default: "^output")}
}
\value{
Nothing
}
\description{
Generate diagnostic plots for all samples.
}
