% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum_stage_fastqc.R
\name{dimsum_stage_fastqc}
\alias{dimsum_stage_fastqc}
\title{dimsum_stage_fastqc}
\usage{
dimsum_stage_fastqc(
  dimsum_meta,
  fastqc_outpath,
  report = TRUE,
  report_outpath = NULL,
  save_workspace = TRUE
)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}

\item{fastqc_outpath}{FASTQC output path (required)}

\item{report}{whether or not to generate FASTQC summary plots (default: TRUE)}

\item{report_outpath}{FASTQC report output path}

\item{save_workspace}{whether or not to save the current workspace (default: TRUE)}
}
\value{
an updated experiment metadata object
}
\description{
Run FASTQC on all fastq files.
}
