% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum_stage_merge.R
\name{dimsum_stage_merge}
\alias{dimsum_stage_merge}
\title{dimsum_stage_merge}
\usage{
dimsum_stage_merge(dimsum_meta, merge_outpath, report = TRUE,
  report_outpath = NULL, save_workspace = TRUE)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}

\item{merge_outpath}{merged variant data output path (required)}

\item{report}{whether or not to generate final summary plots (default: TRUE)}

\item{report_outpath}{final summary report output path}

\item{save_workspace}{whether or not to save the current workspace (default: TRUE)}
}
\value{
an updated experiment metadata object
}
\description{
Merge all variant files.
}
