% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum_stage_merge_report.R
\name{dimsum_stage_merge_report}
\alias{dimsum_stage_merge_report}
\title{dimsum_stage_merge_report}
\usage{
dimsum_stage_merge_report(dimsum_meta, report_outpath)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}

\item{report_outpath}{final report output path (required)}
}
\value{
a single data.frame where empty rows are filled with NAs
}
\description{
Generate final summary plots for all samples.
}
