% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum_stage_split.R
\name{dimsum_stage_split}
\alias{dimsum_stage_split}
\title{dimsum_stage_split}
\usage{
dimsum_stage_split(dimsum_meta, split_outpath, execute = TRUE,
  save_workspace = TRUE)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}

\item{split_outpath}{split FASTQ output path (required)}

\item{execute}{whether or not to execute the system command (default: TRUE)}

\item{save_workspace}{whether or not to save the current experiment metadata object (default: TRUE)}
}
\value{
an updated experiment metadata object
}
\description{
Split all fastq files.
}
