% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum_stage_unique.R
\name{dimsum_stage_unique}
\alias{dimsum_stage_unique}
\title{dimsum_stage_unique}
\usage{
dimsum_stage_unique(dimsum_meta, unique_outpath, save_workspace = TRUE)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}

\item{unique_outpath}{fastx_collapser output path (required)}

\item{save_workspace}{whether or not to save the current experiment metadata object (default: TRUE)}
}
\value{
an updated experiment metadata object
}
\description{
Run fastx_collapser on all fastq files.
}
