% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum_stage_unzip.R
\name{dimsum_stage_unzip}
\alias{dimsum_stage_unzip}
\title{dimsum_stage_unzip}
\usage{
dimsum_stage_unzip(dimsum_meta, fastq_outpath, save_workspace = TRUE)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}

\item{fastq_outpath}{FASTQ output path (required)}

\item{save_workspace}{whether or not to save the current workspace (default: TRUE)}
}
\value{
an updated experiment metadata object
}
\description{
Unzip all fastq files.
}
