% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum_stage_usearch.R
\name{dimsum_stage_usearch}
\alias{dimsum_stage_usearch}
\title{dimsum_stage_usearch}
\usage{
dimsum_stage_usearch(dimsum_meta, usearch_outpath, execute = TRUE,
  report = TRUE, report_outpath = NULL, save_workspace = TRUE)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}

\item{usearch_outpath}{USEARCH output path (required)}

\item{execute}{whether or not to execute the system command (default: TRUE)}

\item{report}{whether or not to generate USEARCH summary plots (default: TRUE)}

\item{report_outpath}{USEARCH report output path}

\item{save_workspace}{whether or not to save the current experiment metadata object (default: TRUE)}
}
\value{
an updated experiment metadata object
}
\description{
Run USEARCH on all fastq files.
}
