% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum_stage_usearch_report.R
\name{dimsum_stage_usearch_report}
\alias{dimsum_stage_usearch_report}
\title{dimsum_stage_usearch_report}
\usage{
dimsum_stage_usearch_report(dimsum_meta, report_outpath)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}

\item{report_outpath}{USEARCH report output path (required)}
}
\value{
an updated experiment metadata object
}
\description{
Generate USEARCH summary plots for all samples.
}
