% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum_stage_vsearch.R
\name{dimsum_stage_vsearch}
\alias{dimsum_stage_vsearch}
\title{dimsum_stage_vsearch}
\usage{
dimsum_stage_vsearch(
  dimsum_meta,
  vsearch_outpath,
  report = TRUE,
  report_outpath = NULL,
  save_workspace = TRUE
)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}

\item{vsearch_outpath}{VSEARCH output path (required)}

\item{report}{whether or not to generate VSEARCH summary plots (default: TRUE)}

\item{report_outpath}{VSEARCH report output path}

\item{save_workspace}{whether or not to save the current workspace (default: TRUE)}
}
\value{
an updated experiment metadata object
}
\description{
Run VSEARCH on all fastq files.
}
