% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastq_manualalign.R
\name{fastq_manualalign}
\alias{fastq_manualalign}
\title{fastq_manualalign}
\usage{
fastq_manualalign(input_FASTQ1, input_FASTQ2, output_FASTQ, output_REPORT,
  num_nuc, min_qual, max_ee, min_len, concatentate_reads = FALSE)
}
\arguments{
\item{input_FASTQ1}{Path to first read FASTQ file (required)}

\item{input_FASTQ2}{Path to second read FASTQ file (required)}

\item{output_FASTQ}{Path to output FASTQ file (required)}

\item{output_REPORT}{Path to report file (required)}

\item{num_nuc}{Enforced alignment length in base pairs (required)}

\item{min_qual}{Minimum observed base quality to retain read pair (required)}

\item{max_ee}{Maximum number of expected errors to retain read pair (required)}

\item{min_len}{Discard pair if either read is shorter than this (required)}

\item{concatentate_reads}{Simply concatentate reads (without reverse complementing second read in pair) (default: FALSE)}
}
\value{
Nothing
}
\description{
Attempt to merge paired end reads according to a specified alignment length.
}
