% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastq_splitter.R
\name{fastq_splitter}
\alias{fastq_splitter}
\title{fastq_splitter}
\usage{
fastq_splitter(inputFile, outputFilePrefix, chunkSize = NULL,
  numRecords = NULL)
}
\arguments{
\item{inputFile}{Path to FASTQ file (required)}

\item{outputFilePrefix}{Prefix to output FASTQ file (required)}

\item{chunkSize}{Chunk size in bytes}

\item{numRecords}{Number of FASTQ records per file}
}
\value{
Number of FASTQ records per file
}
\description{
Split FASTQ file into roughly equally sized chunks (bytes).
}
