% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastq_splitter.R
\name{fastq_splitter_writeFastq}
\alias{fastq_splitter_writeFastq}
\title{fastq_splitter_writeFastq}
\usage{
fastq_splitter_writeFastq(shortreads, outputFile, initial_write)
}
\arguments{
\item{shortreads}{ShortReadQ object (required)}

\item{outputFile}{Path to output FASTQ file (required)}

\item{initial_write}{Is this the first write? (required)}
}
\value{
Nothing
}
\description{
writeFastq wrapper for fastq_splitter.
}
