% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_experiment_design.R
\name{get_experiment_design}
\alias{get_experiment_design}
\title{get_experiment_design}
\usage{
get_experiment_design(dimsum_meta)
}
\arguments{
\item{dimsum_meta}{path to experiment design file (required)}
}
\value{
a data.frame
}
\description{
Get and format metadata from experiment design file.
}
