% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_cutadapt_output.R
\name{parse_cutadapt_output}
\alias{parse_cutadapt_output}
\title{parse_cutadapt_output}
\usage{
parse_cutadapt_output(file_path)
}
\arguments{
\item{file_path}{path to cutadapt output file (required)}
}
\value{
a list of read trimming count statistics
}
\description{
Parse cutadapt stdout.
}
