% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports_summary.R
\name{reports_summary}
\alias{reports_summary}
\title{reports_summary}
\usage{
reports_summary(dimsum_meta)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}
}
\value{
a summary report character string (html format)
}
\description{
Generate a summary report in html format.
}
