% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_count_distributions_by_ntmut.R
\name{sample_count_distributions_by_ntmut}
\alias{sample_count_distributions_by_ntmut}
\title{sample_count_distributions_by_ntmut}
\usage{
sample_count_distributions_by_ntmut(input_dt, output_file, width = 12,
  height = 8, title = "")
}
\arguments{
\item{input_dt}{input data table (required)}

\item{output_file}{plot output path (required)}

\item{width}{plot width (default: 12)}

\item{height}{plot height (default: 8)}

\item{title}{plot title (default: "")}
}
\value{
Nothing
}
\description{
Plot sample sample count distributions split by number of nucleotide mutations (Nmut_nt) and synonymous/nonsynonymous mutations (Nmut_aa).
}
