% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum.R
\name{dimsum}
\alias{dimsum}
\title{Run DiMSum pipeline}
\usage{
dimsum(fastqFileDir, fastqFileExtension = ".fastq", gzipped = T,
  stranded = T, paired = T, barcodeDesignPath,
  barcodeErrorRate = 0.25, experimentDesignPath, cutadaptCut5First,
  cutadaptCut5Second, cutadaptCut3First, cutadaptCut3Second,
  cutadapt5First, cutadapt5Second, cutadapt3First, cutadapt3Second,
  cutadaptMinLength = 50, cutadaptErrorRate = 0.2, usearchMinQual,
  usearchMaxee, usearchMinlen = 64, usearchMinovlen = 16, outputPath,
  projectName, wildtypeSequence, sequenceType = "auto",
  transLibrary = F, bayesianDoubleFitness = F,
  bayesianDoubleFitnessLamD = 0.025, fitnessMinInputCountAll = 0,
  fitnessMinInputCountAny = 5, fitnessHighConfidenceCount = 10,
  fitnessDoubleHighConfidenceCount = 50, fitnessMaxSubstitutions = 2,
  retainIntermediateFiles = F, splitChunkSize = 3758096384,
  retainedReplicates = "all", startStage = 1, stopStage = 0,
  numCores = 1)
}
\arguments{
\item{fastqFileDir}{Path to directory with input FASTQ files}

\item{fastqFileExtension}{FASTQ file extension (default:".fastq")}

\item{gzipped}{Are FASTQ files are gzipped? (default:T)}

\item{stranded}{Is the library design stranded? (default:T)}

\item{paired}{Is the library design paired-end? (default:T)}

\item{barcodeDesignPath}{Path to barcode design file (tab-separated plain text file with barcode design)}

\item{barcodeErrorRate}{Maximum allowed error rate for the barcode (default:0.25)}

\item{experimentDesignPath}{Path to experimental design file (tab-separated plain text file with replicate structure)}

\item{cutadaptCut5First}{cutadapt: remove bases from start of first read (before constant region trimming)}

\item{cutadaptCut5Second}{cutadapt: remove bases from start of second read (before constant region trimming)}

\item{cutadaptCut3First}{cutadapt: remove bases from end of first read (before constant region trimming)}

\item{cutadaptCut3Second}{cutadapt: remove bases from end of second read (before constant region trimming)}

\item{cutadapt5First}{cutadapt: sequence of 5' constant region (of the first read)}

\item{cutadapt5Second}{cutadapt: sequence of 5' constant region (of the second read)}

\item{cutadapt3First}{cutadapt: sequence of 3' constant region (of the first read; default: reverse complement of cutadapt5Second)}

\item{cutadapt3Second}{cutadapt: sequence of 3' constant region (of the second read; default: reverse complement of cutadapt5First)}

\item{cutadaptMinLength}{cutadapt: Discard reads shorter than LENGTH (default:50)}

\item{cutadaptErrorRate}{cutadapt: Maximum allowed error rate (default:0.2)}

\item{usearchMinQual}{USEARCH: minimum observed base quality to retain read pair}

\item{usearchMaxee}{USEARCH: maximum number of expected errors to retain read pair}

\item{usearchMinlen}{USEARCH: Discard pair if either read is shorter than this (default:64)}

\item{usearchMinovlen}{USEARCH: discard pair if alignment is shorter than given value (default:16)}

\item{outputPath}{Path to directory to use for output files}

\item{projectName}{Project name}

\item{wildtypeSequence}{Wild-type nucleotide sequence (A/C/G/T). Lower-case letters (a/c/g/t) indicate internal constant regions to be removed before fitness calculations.}

\item{sequenceType}{Coding potential of sequence; either noncoding/coding/auto (default:auto)}

\item{transLibrary}{Trans library design i.e. read pairs correspond to distinct peptides (no overlap)}

\item{bayesianDoubleFitness}{Improve double mutant fitness estimates using Bayesian framework (default:F)}

\item{bayesianDoubleFitnessLamD}{Poisson distribution for score likelihood (default:0.025)}

\item{fitnessMinInputCountAll}{Minimum input read count (in all replicate) to be retained during fitness calculations (default:0)}

\item{fitnessMinInputCountAny}{Minimum input read count (in any replicate) to be retained during fitness calculations (default:5)}

\item{fitnessHighConfidenceCount}{Minimum mean input read count for high confidence variants (default:10)}

\item{fitnessDoubleHighConfidenceCount}{Minimum input replicate read count for doubles used to derive prior for Bayesian doubles correction (default:50)}

\item{fitnessMaxSubstitutions}{Maximum number of nucleotide or amino acid substitutions for coding or non-coding sequences respectively (default:2)}

\item{retainIntermediateFiles}{Should intermediate files be retained? (default:F)}

\item{splitChunkSize}{FASTQ file split chunk size in bytes (default:3758096384)}

\item{retainedReplicates}{Comma-separated list of Input replicates (or experiment ids) to retain or 'all' (default:'all')}

\item{startStage}{Start at a specified pipeline stage (default:1)}

\item{stopStage}{Stop at a specified pipeline stage (default:0 i.e. no stop condition)}

\item{numCores}{Number of available CPU cores (default:1)}
}
\value{
Nothing
}
\description{
This function runs the DiMSum pipeline.
}
