% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum.R
\name{dimsum}
\alias{dimsum}
\title{Run DiMSum pipeline}
\usage{
dimsum(fastqFileDir, fastqFileExtension = ".fastq", gzipped = T,
  stranded = T, paired = T, barcodeDesignPath,
  barcodeErrorRate = 0.25, experimentDesignPath,
  experimentDesignPairDuplicates = F, barcodeIdentityPath, countPath,
  cutadaptCut5First, cutadaptCut5Second, cutadaptCut3First,
  cutadaptCut3Second, cutadapt5First, cutadapt5Second, cutadapt3First,
  cutadapt3Second, cutadaptMinLength = 50, cutadaptErrorRate = 0.2,
  cutadaptOverlap = 3, usearchMinQual = 30, usearchMaxee = 0.5,
  usearchMinlen = 64, usearchMinovlen = 16, outputPath, projectName,
  wildtypeSequence, permittedSequences, reverseComplement = F,
  sequenceType = "auto", mutagenesisType = "random",
  transLibrary = F, transLibraryReverseComplement = F,
  bayesianDoubleFitness = F, bayesianDoubleFitnessLamD = 0.025,
  fitnessMinInputCountAll = 0, fitnessMinInputCountAny = 0,
  fitnessMinOutputCountAll = 0, fitnessMinOutputCountAny = 0,
  fitnessHighConfidenceCount = 10,
  fitnessDoubleHighConfidenceCount = 50, fitnessNormalise = T,
  fitnessErrorModel = T, maxSubstitutions = 2,
  mixedSubstitutions = F, retainIntermediateFiles = F,
  splitChunkSize = 3758096384, retainedReplicates = "all",
  startStage = 0, stopStage = 5, numCores = 1)
}
\arguments{
\item{fastqFileDir}{Path to directory with input FASTQ files (required for WRAP)}

\item{fastqFileExtension}{FASTQ file extension (default:'.fastq')}

\item{gzipped}{Are FASTQ files are gzipped? (default:T)}

\item{stranded}{Is the library design stranded? (default:T)}

\item{paired}{Is the library design paired-end? (default:T)}

\item{barcodeDesignPath}{Path to barcode design file (tab-separated plain text file with barcode design)}

\item{barcodeErrorRate}{Maximum allowed error rate for the barcode (default:0.25)}

\item{experimentDesignPath}{Path to experimental design file (tab-separated plain text file with replicate structure)}

\item{experimentDesignPairDuplicates}{Are duplicate FASTQ files permitted in experimental design file? (default:F)}

\item{barcodeIdentityPath}{Path to barcode identity file (tab-separated plain text file mapping barcodes to variants)}

\item{countPath}{Path to variant count file for analysis with STEAM only (tab-separated plain text file with sample counts for all variants)}

\item{cutadaptCut5First}{cutadapt: remove bases from start of first read (before constant region trimming)}

\item{cutadaptCut5Second}{cutadapt: remove bases from start of second read (before constant region trimming)}

\item{cutadaptCut3First}{cutadapt: remove bases from end of first read (before constant region trimming)}

\item{cutadaptCut3Second}{cutadapt: remove bases from end of second read (before constant region trimming)}

\item{cutadapt5First}{cutadapt: sequence of 5' constant region to be trimmed (of the first read)}

\item{cutadapt5Second}{cutadapt: sequence of 5' constant region to be trimmed (of the second read)}

\item{cutadapt3First}{cutadapt: sequence of 3' constant region to be trimmed (of the first read; default: reverse complement of cutadapt5Second if not linked)}

\item{cutadapt3Second}{cutadapt: sequence of 3' constant region to be trimmed (of the second read; default: reverse complement of cutadapt5First if not linked)}

\item{cutadaptMinLength}{cutadapt: Discard reads shorter than LENGTH after trimming (default:50)}

\item{cutadaptErrorRate}{cutadapt: Maximum allowed error rate for trimming (default:0.2)}

\item{cutadaptOverlap}{cutadapt: Minimum overlap between read and constant region for trimming (default:3)}

\item{usearchMinQual}{USEARCH: minimum observed base quality to retain read pair (default:30)}

\item{usearchMaxee}{USEARCH: maximum number of expected errors to retain read pair (default:0.5)}

\item{usearchMinlen}{USEARCH: Discard pair if either read is shorter than this (default:64)}

\item{usearchMinovlen}{USEARCH: discard pair if alignment is shorter than given value (default:16)}

\item{outputPath}{Path to directory to use for output files}

\item{projectName}{Project name}

\item{wildtypeSequence}{Wild-type nucleotide sequence (A/C/G/T). Lower-case letters (a/c/g/t) indicate internal constant regions to be removed during WRAP.}

\item{permittedSequences}{A sequence of nucleotide codes (A/C/G/T/R/Y/S/W/K/M/B/D/H/V/N) with length matching the number of mutated positions i.e upper-case letters in wild-type nucleotide sequence (default:any base at mutated positions)}

\item{reverseComplement}{Reverse complement variant sequences before processing? (default:F)}

\item{sequenceType}{Coding potential of sequence; either noncoding/coding/auto (default:auto)}

\item{mutagenesisType}{Whether mutagenesis was performed at the nucleotide or codon/amino acid level; either random/codon (default:random)}

\item{transLibrary}{Trans library design i.e. read pairs correspond to distinct peptides with no overlap (default:F)}

\item{transLibraryReverseComplement}{Reverse complement second read in pair before concatenating (default:F)}

\item{bayesianDoubleFitness}{Improve double mutant fitness estimates using Bayesian framework (DISABLED: still in development)}

\item{bayesianDoubleFitnessLamD}{Poisson distribution for score likelihood (default:0.025)}

\item{fitnessMinInputCountAll}{Minimum input read count (in all replicate) to be retained during fitness calculations (default:0)}

\item{fitnessMinInputCountAny}{Minimum input read count (in any replicate) to be retained during fitness calculations (default:0)}

\item{fitnessMinOutputCountAll}{Minimum output read count (in all replicate) to be retained during fitness calculations (default:0)}

\item{fitnessMinOutputCountAny}{Minimum output read count (in any replicate) to be retained during fitness calculations (default:0)}

\item{fitnessHighConfidenceCount}{Minimum mean input read count for high confidence variants (default:10)}

\item{fitnessDoubleHighConfidenceCount}{Minimum input replicate read count for doubles used to derive prior for Bayesian doubles correction (default:50)}

\item{fitnessNormalise}{Normalise fitness values to minimise inter-replicate differences (default:T)}

\item{fitnessErrorModel}{Fit fitness error model (default:T)}

\item{maxSubstitutions}{Maximum number of nucleotide or amino acid substitutions for coding or non-coding sequences respectively (default:2)}

\item{mixedSubstitutions}{For coding sequences, are nonsynonymous variants with silent/synonymous substitutions in other codons allowed? (default:F)}

\item{retainIntermediateFiles}{Should intermediate files be retained? (default:F)}

\item{splitChunkSize}{FASTQ file split chunk size in bytes (default:3758096384)}

\item{retainedReplicates}{Comma-separated list of Input replicates (or experiment ids) to retain or 'all' (default:'all')}

\item{startStage}{Start at a specified pipeline stage (default:0)}

\item{stopStage}{Stop at a specified pipeline stage (default:5)}

\item{numCores}{Number of available CPU cores (default:1)}
}
\value{
Nothing
}
\description{
This function runs the DiMSum pipeline.
}
