% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__aggregate_AA_variants.R
\name{dimsum__aggregate_AA_variants}
\alias{dimsum__aggregate_AA_variants}
\title{dimsum__aggregate_AA_variants}
\usage{
dimsum__aggregate_AA_variants(dimsum_meta, input_dt, all_reps)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}

\item{input_dt}{input data.table (required)}

\item{all_reps}{list of replicates to retain (required)}
}
\value{
data.table with aggregated counts
}
\description{
Aggregate counts from variants that are identical at the AA level and without synonymous mutations.
}
