% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__bayesian_double_fitness.R
\name{dimsum__bayesian_double_fitness}
\alias{dimsum__bayesian_double_fitness}
\title{dimsum__bayesian_double_fitness}
\usage{
dimsum__bayesian_double_fitness(
  dimsum_meta,
  doubles_dt,
  singles_dt,
  wt_dt,
  all_reps,
  report = TRUE,
  report_outpath = NULL
)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}

\item{doubles_dt}{doubles data.table (required)}

\item{singles_dt}{singles data.table (required)}

\item{wt_dt}{WT data.table (required)}

\item{all_reps}{list of replicates to retain (required)}

\item{report}{whether or not to generate fitness summary plots (default: TRUE)}

\item{report_outpath}{fitness report output path}
}
\value{
Nothing
}
\description{
Calculating Bayesian double mutant fitness estimates.
}
