% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__calculate_fitness.R
\name{dimsum__calculate_fitness}
\alias{dimsum__calculate_fitness}
\title{dimsum__calculate_fitness}
\usage{
dimsum__calculate_fitness(input_dt, all_reps, sequence_type,
  report = TRUE, report_outpath = NULL, verbose = T)
}
\arguments{
\item{input_dt}{input data.table (required)}

\item{all_reps}{list of replicates to retain (required)}

\item{sequence_type}{coding potential of sequence: noncoding/coding (required)}

\item{report}{whether or not to generate fitness summary plots (default: TRUE)}

\item{report_outpath}{fitness report output path}

\item{verbose}{whether or not to print status messages (default: TRUE)}
}
\value{
Nothing
}
\description{
Calculate fitness and count-based error.
}
