% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__check_countfile.R
\name{dimsum__check_countfile}
\alias{dimsum__check_countfile}
\title{dimsum__check_countfile}
\usage{
dimsum__check_countfile(dimsum_meta, input_dt)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}

\item{input_dt}{input data.table (required)}
}
\value{
Reformatted data.table
}
\description{
Check whether user-specified count file correctly formatted.
}
