% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__check_experiment_design.R
\name{dimsum__check_experiment_design}
\alias{dimsum__check_experiment_design}
\title{dimsum__check_experiment_design}
\usage{
dimsum__check_experiment_design(exp_design)
}
\arguments{
\item{exp_design}{experiment design data.frame (required)}
}
\value{
Nothing
}
\description{
Validate metadata from experiment design file.
}
