% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__concatenate_reads.R
\name{dimsum__concatenate_reads}
\alias{dimsum__concatenate_reads}
\title{dimsum__concatenate_reads}
\usage{
dimsum__concatenate_reads(input_FASTQ1, input_FASTQ2, output_FASTQ,
  output_REPORT, min_qual, max_ee, min_len)
}
\arguments{
\item{input_FASTQ1}{Path to first read FASTQ file (required)}

\item{input_FASTQ2}{Path to second read FASTQ file (required)}

\item{output_FASTQ}{Path to output FASTQ file (required)}

\item{output_REPORT}{Path to report file (required)}

\item{min_qual}{Minimum observed base quality to retain read pair (required)}

\item{max_ee}{Maximum number of expected errors to retain read pair (required)}

\item{min_len}{Discard pair if either read is shorter than this (required)}
}
\value{
Nothing
}
\description{
Concatentate reads (without reverse complementing second read in pair).
}
