% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__debarcode_variants.R
\name{dimsum__debarcode_variants}
\alias{dimsum__debarcode_variants}
\title{dimsum__debarcode_variants}
\usage{
dimsum__debarcode_variants(dimsum_meta, input_dt)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}

\item{input_dt}{input data.table (required)}
}
\value{
A data.table with variants debarcoded
}
\description{
Debarcode variants.
}
