% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__filter_nuc_variants.R
\name{dimsum__filter_nuc_variants}
\alias{dimsum__filter_nuc_variants}
\title{dimsum__filter_nuc_variants}
\usage{
dimsum__filter_nuc_variants(dimsum_meta, input_dt, all_reps)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}

\item{input_dt}{output path for plots and saved objects (required)}

\item{all_reps}{list of replicates to retain (required)}
}
\value{
Nothing
}
\description{
Filter out low count nucleotide variants.
}
