% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__filter_variants_coding.R
\name{dimsum__filter_variants_coding}
\alias{dimsum__filter_variants_coding}
\title{dimsum__filter_variants_coding}
\usage{
dimsum__filter_variants_coding(dimsum_meta, input_dt, wt_ntseq, all_reps,
  report = TRUE, report_outpath = NULL)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}

\item{input_dt}{output path for plots and saved objects (required)}

\item{wt_ntseq}{WT nucleotide sequence (required)}

\item{all_reps}{list of replicates to retain (required)}

\item{report}{whether or not to generate fitness summary plots (default: TRUE)}

\item{report_outpath}{fitness report output path}
}
\value{
Nothing
}
\description{
Filter for desired coding variants.
}
