% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__identify_double_aa_mutations.R
\name{dimsum__identify_double_aa_mutations}
\alias{dimsum__identify_double_aa_mutations}
\title{dimsum__identify_double_aa_mutations}
\usage{
dimsum__identify_double_aa_mutations(input_dt, singles_dt, wt_AAseq)
}
\arguments{
\item{input_dt}{input data.table (required)}

\item{singles_dt}{singles data.table (required)}

\item{wt_AAseq}{WT amino acid sequence (required)}
}
\value{
data.table with double amino acid variants
}
\description{
Identify and annotate double AA substitutions.
}
