% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__identify_single_nt_mutations.R
\name{dimsum__identify_single_nt_mutations}
\alias{dimsum__identify_single_nt_mutations}
\title{dimsum__identify_single_nt_mutations}
\usage{
dimsum__identify_single_nt_mutations(input_dt, wt_ntseq, report = TRUE,
  report_outpath = NULL)
}
\arguments{
\item{input_dt}{input data.table (required)}

\item{wt_ntseq}{WT nucleotide sequence (required)}

\item{report}{whether or not to generate fitness summary plots (default: TRUE)}

\item{report_outpath}{fitness report output path}
}
\value{
Nothing
}
\description{
Identify and annotate single nucleotide substitutions.
}
