% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__parse_cutadapt_output.R
\name{dimsum__parse_cutadapt_output}
\alias{dimsum__parse_cutadapt_output}
\title{dimsum__parse_cutadapt_output}
\usage{
dimsum__parse_cutadapt_output(file_path, ran_cutadapt = T)
}
\arguments{
\item{file_path}{path to cutadapt output file (required)}

\item{ran_cutadapt}{whether this is a bona fide cutadapt output file (default:T)}
}
\value{
a list of read trimming count statistics
}
\description{
Parse cutadapt stdout.
}
