% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__parse_cutadapt_output_single_end.R
\name{dimsum__parse_cutadapt_output_single_end}
\alias{dimsum__parse_cutadapt_output_single_end}
\title{dimsum__parse_cutadapt_output_single_end}
\usage{
dimsum__parse_cutadapt_output_single_end(file_path, ran_cutadapt = T)
}
\arguments{
\item{file_path}{path to cutadapt output file (required)}

\item{ran_cutadapt}{whether this is a bona fide cutadapt output file (default:T)}
}
\value{
a list of read trimming count statistics
}
\description{
Parse cutadapt stdout (single-end mode).
}
