% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__remove_internal_constant_region.R
\name{dimsum__remove_internal_constant_region}
\alias{dimsum__remove_internal_constant_region}
\title{dimsum__remove_internal_constant_region}
\usage{
dimsum__remove_internal_constant_region(input_dt, wt_ccntseq)
}
\arguments{
\item{input_dt}{output path for plots and saved objects (required)}

\item{wt_ccntseq}{case-coded WT nucleotide sequence (required)}
}
\value{
A data.table with internal constant regions removed and
}
\description{
Remove internal constant region from nucleotide sequences (if perfect match found).
}
