% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum__status_message.R
\name{dimsum__status_message}
\alias{dimsum__status_message}
\title{dimsum__status_message}
\usage{
dimsum__status_message(..., newline = F)
}
\arguments{
\item{...}{objects to be output as character vector (required)}

\item{newline}{print newline after status message (default:F)}
}
\value{
Nothing
}
\description{
Write status message to standard output.
}
