% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum_stage_counts_to_fitness.R
\name{dimsum_stage_counts_to_fitness}
\alias{dimsum_stage_counts_to_fitness}
\title{dimsum_stage_counts_to_fitness}
\usage{
dimsum_stage_counts_to_fitness(
  dimsum_meta,
  fitness_outpath,
  report = TRUE,
  report_outpath = NULL,
  save_workspace = TRUE
)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}

\item{fitness_outpath}{output path for saved objects (required)}

\item{report}{whether or not to generate fitness summary plots (default: TRUE)}

\item{report_outpath}{fitness report output path}

\item{save_workspace}{whether or not to save the current workspace (default: TRUE)}
}
\value{
Nothing
}
\description{
Estimate fitness of singles, doubles and silent AA substitution variants from DiMSum formatted output counts.
}
