% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsum_stage_diagnostics_report.R
\name{dimsum_stage_diagnostics_report}
\alias{dimsum_stage_diagnostics_report}
\title{dimsum_stage_diagnostics_report}
\usage{
dimsum_stage_diagnostics_report(dimsum_meta, report_outpath,
  input_samples_pattern = "_s0_", output_samples_pattern = "_s1_")
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}

\item{report_outpath}{diagnostics report output path (required)}

\item{input_samples_pattern}{input samples string pattern (default: "_s0_")}

\item{output_samples_pattern}{diagnostics report output path (default: "_s1_")}
}
\value{
Nothing
}
\description{
Generate diagnostic plots for all samples.
}
