% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_metadata.R
\name{save_metadata}
\alias{save_metadata}
\title{save_metadata}
\usage{
save_metadata(dimsum_meta, n)
}
\arguments{
\item{dimsum_meta}{an experiment metadata object (required)}

\item{n}{the number of generations to go back (required)}
}
\value{
Nothing
}
\description{
Save experiment workspace including metadata and session information.
}
